Add fallback RPC support on the frontend
The wagmi config built its transports with a bare `http()`, so only
the first RPC URL was ever used and a failing endpoint degraded the
whole app. Transports now come from a shared `createRpcTransport`,
which wraps several URLs in a viem `fallback` so a dead or
rate-limited endpoint rolls over to the next.

`updateRpcUrls` had three copies (web, api, internal-dashboard) and
`createMainnetClient` lived in the api. Both now live in
`@vetro-protocol/core` next to `createRpcTransport`, so all three
apps share the same fallback, request batching and Multicall3
config.

Passing several URLs stays a deploy-time decision: `.env` keeps a
single URL so the anvil fork used by e2e and `dev:fork` still gets
a usable `--fork-url`, and the `+`-joined list is set in the
Cloudflare build config.
