+++ api/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@sentry/cloudflare": "10.58.0",
+    "@vetro-protocol/core": "workspace:*",
     "@vetro-protocol/earn": "workspace:*",
     "@vetro-protocol/gateway": "workspace:*",
     "@vetro-protocol/treasury": "workspace:*",
