+++ api/src/total-deposits-history.ts
@@ -1,8 +1,8 @@
+import { createMainnetClient } from "@vetro-protocol/core";
 import { type Address } from "viem";
 import { totalAssets } from "viem-erc4626/actions";
 
 import { appendLivePoint } from "./append-live-point.ts";
-import { createMainnetClient } from "./mainnet-client.ts";
 import { paginateSubgraphQuery } from "./paginate-subgraph-query.ts";
 import { getPeriodStart } from "./vault-history-period.ts";
