+++ packages/core/src/createMainnetClient.ts
@@ -1,0 +1,19 @@
+import { createPublicClient } from "viem";
+import { mainnet } from "viem/chains";
+
+import { createRpcTransport } from "./createRpcTransport.ts";
+import { updateRpcUrls } from "./utils/updateRpcUrls.ts";
+
+/**
+ * Create a viem public client for Ethereum mainnet from the given RPC URL
+ * config. See `updateRpcUrls` for the accepted formats, and
+ * `createRpcTransport` for how they become a transport.
+ */
+export const createMainnetClient = function (rpcUrl: string | undefined) {
+  const chain = updateRpcUrls(mainnet, rpcUrl);
+  return createPublicClient({
+    batch: { multicall: true },
+    chain,
+    transport: createRpcTransport(chain),
+  });
+};
