+++ packages/core/src/index.ts
@@ -1,3 +1,5 @@
+export { createMainnetClient } from "./createMainnetClient.ts";
+export { createRpcTransport } from "./createRpcTransport.ts";
 export {
   sVetBtcAddress,
   sVusdAddress,
@@ -6,3 +8,4 @@ export {
 export { knownTokens } from "./tokens.ts";
 export type { Token } from "./types.ts";
 export { isAddressValid } from "./utils/isAddressValid.ts";
+export { updateRpcUrls } from "./utils/updateRpcUrls.ts";
